requirements.txt
streamlit
opencv-python
numpy
Pillow
torch
torchvision
seaborn
timm==0.6.7
ultralytics